% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galois_field.R
\name{create_galois_field}
\alias{create_galois_field}
\title{Create a Galois field}
\usage{
create_galois_field(q)
}
\arguments{
\item{q}{The order of the Galois Field q = p^n}
}
\value{
a GaloisField object
}
\description{
Create a Galois field
}
\examples{
gf <- create_galois_field(4);
}
