% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_oalhs.R
\name{create_oalhs}
\alias{create_oalhs}
\title{Create an orthogonal array Latin hypercube}
\usage{
create_oalhs(n, k, bChooseLargerDesign, bverbose)
}
\arguments{
\item{n}{the number of samples or rows in the LHS (integer)}

\item{k}{the number of parameters or columns in the LHS (integer)}

\item{bChooseLargerDesign}{should a larger oa design be chosen than the n and k requested?}

\item{bverbose}{should information be printed with execution}
}
\value{
a numeric matrix which is an orthogonal array Latin hypercube sample
}
\description{
Create an orthogonal array Latin hypercube
}
\examples{
create_oalhs(9, 4, TRUE, TRUE)
}
