% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOA.R
\name{createAddelKemp}
\alias{createAddelKemp}
\title{Create an orthogonal array using the Addelman-Kempthorne algorithm}
\usage{
createAddelKemp(q, ncol, bRandom)
}
\arguments{
\item{q}{the number of symbols in the array}

\item{ncol}{number of parameters or columns}

\item{bRandom}{should the array be randomized}
}
\value{
an orthogonal array
}
\description{
Create an orthogonal array using the Addelman-Kempthorne algorithm
}
\examples{
A <- createAddelKemp(3, 3, TRUE)
B <- createAddelKemp(3, 5, FALSE)
}
\seealso{
Other methods to create orthogonal arrays [createBoseBush()],
[createBose()], [createAddelKemp3()],
[createAddelKempN()], [createBusht()], [createBoseBushl()]
}
