% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOA.R
\name{createAddelKemp3}
\alias{createAddelKemp3}
\title{Create an orthogonal array using the Addelman-Kempthorne algorithm
with 2*q^3 rows}
\usage{
createAddelKemp3(q, ncol, bRandom)
}
\arguments{
\item{q}{the number of symbols in the array}

\item{ncol}{number of parameters or columns}

\item{bRandom}{should the array be randomized}
}
\value{
an orthogonal array
}
\description{
Create an orthogonal array using the Addelman-Kempthorne algorithm
with 2*q^3 rows
}
\examples{
A <- createAddelKemp3(3, 3, TRUE)
B <- createAddelKemp3(3, 5, FALSE)
}
\seealso{
Other methods to create orthogonal arrays [createBushBush()],
[createBose()], [createAddelKemp()],
[createAddelKempN()], [createBusht()], [createBoseBushl()]
}
