% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOA.R
\name{createBose}
\alias{createBose}
\title{Create an orthogonal array using the Bose algorithm}
\usage{
createBose(q, ncol, bRandom = TRUE)
}
\arguments{
\item{q}{the number of symbols in the array}

\item{ncol}{number of parameters or columns}

\item{bRandom}{should the array be randomized}
}
\value{
an orthogonal array
}
\description{
Create an orthogonal array using the Bose algorithm
}
\examples{
A <- createBose(3, 3, FALSE)
B <- createBose(5, 4, TRUE)
}
\seealso{
Other methods to create orthogonal arrays [createBush()],
[createBoseBush()], [createAddelKemp()], [createAddelKemp3()],
[createAddelKempN()], [createBusht()], [createBoseBushl()]
}
