% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOA.R
\name{createBoseBush}
\alias{createBoseBush}
\title{Create an orthogonal array using the Bose-Bush algorithm}
\usage{
createBoseBush(q, ncol, bRandom = TRUE)
}
\arguments{
\item{q}{the number of symbols in the array}

\item{ncol}{number of parameters or columns}

\item{bRandom}{should the array be randomized}
}
\value{
an orthogonal array
}
\description{
Create an orthogonal array using the Bose-Bush algorithm
}
\examples{
A <- createBoseBush(4, 3, FALSE)
B <- createBoseBush(8, 3, TRUE)
}
\seealso{
Other methods to create orthogonal arrays [createBush()],
[createBose()], [createAddelKemp()], [createAddelKemp3()],
[createAddelKempN()], [createBusht()], [createBoseBushl()]
}
