% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOA.R
\name{createBush}
\alias{createBush}
\title{Create an orthogonal array using the Bush algorithm}
\usage{
createBush(q, ncol, bRandom = TRUE)
}
\arguments{
\item{q}{the number of symbols in the array}

\item{ncol}{number of parameters or columns}

\item{bRandom}{should the array be randomized}
}
\value{
an orthogonal array
}
\description{
Create an orthogonal array using the Bush algorithm
}
\examples{
A <- createBush(3, 3, FALSE)
B <- createBush(4, 5, TRUE)
}
\seealso{
Other methods to create orthogonal arrays [createBoseBush()],
[createBose()], [createAddelKemp()], [createAddelKemp3()],
[createAddelKempN()], [createBusht()], [createBoseBushl()]
}
