% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_library_versions.R
\name{get_library_versions}
\alias{get_library_versions}
\title{Get version information for all libraries in the lhs package}
\usage{
get_library_versions()
}
\value{
a character string containing the versions
}
\description{
Get version information for all libraries in the lhs package
}
\examples{
get_library_versions()
}
