% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOA.R
\name{getLibraryVersions}
\alias{getLibraryVersions}
\title{Get the versions of the underlying bclib, oa, and lhslib C++ libraries}
\usage{
getLibraryVersions()
}
\value{
String of version information
}
\description{
Get the versions of the underlying bclib, oa, and lhslib C++ libraries
}
\examples{
getLibraryVersions()
}
