\name{lhs-internal}
\alias{utility}
\alias{runifint}

\title{
  Internal lhs Functions
}
\description{
  Internal functions for package lhs
}
\details{
  These are not to be called by the user but are used internally in the lhs package.
}
\usage{
  runifint(n=1, min_int=0, max_int=1)
}
\arguments{
  \item{n}{The number of samples}
  \item{min_int}{The lower bound of the distribution (inclusive)}
  \item{max_int}{The upper bound of the distribution (inclusive)}
}
\keyword{
 internal 
}
