% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs.R
\docType{package}
\name{lhs-package}
\alias{lhs}
\alias{lhs-package}
\title{lhs: Latin Hypercube Samples}
\description{
Provides a number of methods for creating and augmenting Latin Hypercube Samples and Orthogonal Array Latin Hypercube Samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bertcarnell/lhs}
  \item Report bugs at \url{https://github.com/bertcarnell/lhs/issues}
}

}
\author{
\strong{Maintainer}: Rob Carnell \email{bertcarnell@gmail.com} (\href{https://orcid.org/0009-0009-0465-7564}{ORCID})

}
\keyword{internal}
