% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_to_oalhs.R
\name{oa_to_oalhs}
\alias{oa_to_oalhs}
\title{Create a Latin hypercube from an orthogonal array}
\usage{
oa_to_oalhs(n, k, oa, bverbose = FALSE)
}
\arguments{
\item{n}{the number of samples or rows in the LHS (integer)}

\item{k}{the number of parameters or columns in the LHS (integer)}

\item{oa}{the orthogonal array to be used as the basis for the LHS (matrix of integers) or data.frame of factors}

\item{bverbose}{should information be printed with execution}
}
\value{
a numeric matrix which is a Latin hypercube sample
}
\description{
Create a Latin hypercube from an orthogonal array
}
\examples{
oa <- createBose(3, 4, TRUE)
oa_to_oalhs(9, 4, oa, TRUE)
}
