% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galois_field.R
\name{poly2int}
\alias{poly2int}
\title{Convert polynomial to integer in <code>0..q-1</code>}
\usage{
poly2int(p, n, poly)
}
\arguments{
\item{p}{modulus}

\item{n}{the length of poly}

\item{poly}{the polynomial vector}
}
\value{
an integer
}
\description{
Convert polynomial to integer in <code>0..q-1</code>
}
\examples{
gf <- create_galois_field(4)
stopifnot(poly2int(gf$p, gf$n, c(0, 0)) == 0)
}
