% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galois_field.R
\name{poly_sum}
\alias{poly_sum}
\title{Addition in polynomial representation}
\usage{
poly_sum(p, n, p1, p2)
}
\arguments{
\item{p}{modulus}

\item{n}{length of polynomial 1 and 2}

\item{p1}{polynomial vector 1}

\item{p2}{polynomial vector 2}
}
\value{
the sum of p1 and p2
}
\description{
Addition in polynomial representation
}
\examples{
gf <- create_galois_field(4)
a <- poly_sum(gf$p, gf$n, c(1, 0), c(0, 1))
stopifnot(all(a == c(1, 1)))
}
