% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_transforms.R
\name{q_factor}
\alias{q_factor}
\alias{q_integer}
\alias{q_dirichlet}
\title{Quantile Transformations}
\usage{
q_factor(p, fact)

q_integer(p, a, b)

q_dirichlet(X, alpha)
}
\arguments{
\item{p}{a vector of LHS samples on (0,1)}

\item{fact}{a factor or categorical variable.  Ordered and un-ordered variables are allowed.}

\item{a}{a minimum integer}

\item{b}{a maximum integer}

\item{X}{multiple columns of an LHS sample on (0,1)}

\item{alpha}{Dirichlet distribution parameters.  $\alpha > 1$}
}
\value{
the transformed column or columns
}
\description{
A collection of functions that transform the margins of a Latin hypercube
sample in non-standard ways
}
\details{
\code{qdirichlet} is not an exact quantile function since the quantile of a
multivariate distribution is not unique.  \code{qdirichlet} is also not the independent quantiles of the marginal distributions since
those quantiles do not sum to one.  \code{qdirichlet} is the quantile of the underlying gamma functions, normalized.
This has been tested to show that \code{qdirichlet} approximates the Dirichlet distribution well and creates the correct marginal means and variances
when using a Latin hypercube sample
}
\examples{
X <- randomLHS(20, 6)
Y <- X
Y[,1] <- qnorm(X[,1], 2, 0.5)
Y[,2] <- q_factor(X[,2], factor(LETTERS[c(1,3,5,7,8)]))
Y[,3] <- q_integer(X[,3], 5, 17)
Y[,4:6] <- q_dirichlet(X[,4:6], c(2,3,4))
}
