% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomLHS.r
\name{randomLHS}
\alias{randomLHS}
\title{Construct a random Latin hypercube design}
\usage{
randomLHS(n, k, preserveDraw = FALSE)
}
\arguments{
\item{n}{the number of rows or samples}

\item{k}{the number of columns or parameters/variables}

\item{preserveDraw}{should the draw be constructed so that it is the same for variable numbers of columns?}
}
\value{
a Latin hypercube sample
}
\description{
\code{randomLHS(4,3)} returns a 4x3 matrix with each column
constructed as follows: A random permutation of (1,2,3,4) is
generated, say (3,1,2,4) for each of K columns.  Then a uniform
random number is picked from each indicated quartile.  In this
example a random number between .5 and .75 is chosen, then one between
0 and .25, then one between .25 and .5, finally one between
.75 and 1.
}
