% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runifint.r
\name{runifint}
\alias{runifint}
\title{Create a Random Sample of Uniform Integers}
\usage{
runifint(n = 1, min_int = 0, max_int = 1)
}
\arguments{
\item{n}{The number of samples}

\item{min_int}{the minimum integer \code{x >= min_int}}

\item{max_int}{the maximum integer \code{x <= max_int}}
}
\value{
the sample sample of size \code{n}
}
\description{
Create a Random Sample of Uniform Integers
}
