% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOA.R
\name{createAddelKempN}
\alias{createAddelKempN}
\title{Create an orthogonal array using the Addelman-Kempthorne algorithm with
alternate strength}
\usage{
createAddelKempN(q, ncol, exponent, bRandom = TRUE)
}
\arguments{
\item{q}{the number of symbols in the array}

\item{ncol}{number of parameters or columns}

\item{exponent}{the exponent on q}

\item{bRandom}{should the array be randomized}
}
\value{
an orthogonal array
}
\description{
Create an orthogonal array using the Addelman-Kempthorne algorithm with
alternate strength
}
\examples{
A <- createAddelKempN(3, 4, 3, TRUE)
B <- createAddelKempN(3, 4, 4, TRUE)
}
\seealso{
Other methods to create orthogonal arrays [createBoseBush()],
[createBose()], [createBush()], [createAddelKemp()], [createAddelKemp3()],
[createBusht()], [createBoseBushl()]
}
