\name{lhs-package}
\alias{lhs-package}
\alias{lhs}
\docType{package}
\title{
The Latin Hypercube Sample (lhs) Package
}
\description{
This package provides a number of methods for creating and augmenting
Latin Hypercube Samples

For a complete list of functions, use library(help="lhs").
}
\author{
Rob Carnell
}
\keyword{design}

