\name{lhs-package}
\alias{lhs-package}
\alias{lhs}
\docType{package}
\title{
The Latin Hypercube Sample (lhs) Package
}
\description{
This package provides a number of methods for creating and augmenting
Latin Hypercube Samples

For a complete list of functions, use library(help="lhs").
}
\author{
Rob Carnell
}
\keyword{design}
\seealso{
  \code{\link{randomLHS}}, \code{\link{geneticLHS}},
  \code{\link{improvedLHS}}, \code{\link{maximinLHS}}, and
  \code{\link{optimumLHS}} to generate Latin Hypercube Samples.
  \code{\link{optAugmentLHS}}, \code{\link{optSeededLHS}}, and
  \code{\link{augmentLHS}} to modify and augment existing designs.
}

